+++ packages/react/src/exampleTypes/Text.js
@@ -1,7 +1,9 @@
 import _ from 'lodash/fp'
 import F from 'futil-js'
 import React from 'react'
-import { contexturify, withTreeLens } from '../utils/hoc'
+import { observer } from 'mobx-react'
+import { withNode, withLoader, withTreeLens } from '../utils/hoc'
+import { withTheme } from '../utils/theme'
 import { setDisplayName } from 'recompose'
 
 let LensInput = ({ lens, theme: { TextInput }, ...props }) => (
@@ -10,8 +12,11 @@ let LensInput = ({ lens, theme: { TextInput }, ...props }) => (
 
 let Text = _.flow(
   setDisplayName('Text'),
+  observer,
   withTreeLens,
-  contexturify
+  withNode,
+  withLoader,
+  withTheme
 )(LensInput)
 
 export default Text