+++ packages/react/src/greyVest/Modal.stories.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import F from 'futil-js'
-import { observer } from 'mobx-react'
 import { useLens } from '../utils/react'
 import { Modal, Button } from '.'
 import decorator from './stories/decorator'
@@ -11,14 +10,24 @@ export default {
   decorators: [decorator],
 }
 
-let ModalDemo = observer(() => {
+export let withOpenProp = () => {
   let open = useLens(false)
   return (
-    <div>
+    <>
       <Modal open={open}>Some Modal Content</Modal>
       <Button onClick={F.on(open)}>Open Modal</Button>
-    </div>
+    </>
   )
-})
+}
 
-export let story = () => <ModalDemo />
+export let with_isOpen_onClose_props = () => {
+  let [isOpen, setIsOpen] = React.useState(false)
+  return (
+    <>
+      <Modal isOpen={isOpen} onClose={() => setIsOpen(false)}>
+        Some Modal Content
+      </Modal>
+      <Button onClick={() => setIsOpen(true)}>Open Modal</Button>
+    </>
+  )
+}