+++ packages/react/src/themes/index.js
@@ -1,4 +1,3 @@
 export base from './base'
 export blueberry from './blueberry'
 export greyVest from './greyVest'
-export material from './material'